Refactor Editor and Landing Components for Improved Functionality and Clarity
- Updated the landing page to change the text from "Transform your screenshots into stunning visuals" to "Transform images into stunning visuals" for better clarity.
- Refactored the EditorLayout component by removing the router and toast notifications, simplifying the user access flow.
- Replaced the image upload logic in the Editor page with a direct check for pro user status, redirecting to the pricing page if necessary.
- Introduced a new EditorPage component to encapsulate image upload and rendering functionality, enhancing code organization and readability.
